feat: 출발지와 도착지 선택 로직 구현
